src/mqtt_kafka_bridge/requirements.txt
paho-mqtt==2.1.0
kafka-python-ng==2.2.2
